requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Bot
python-telegram-bot==20.7

# LLM
openai==1.54.5

# Vector DB
pinecone-client==3.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# File Monitoring
watchdog==3.0.0

# Utils
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1